extract sql parse
